Create separate script entrypoing for SDK
Transcode API had to lose all the logs and inquiries
and moved them directly to the CLI instead.
