+++ bin/video-nft
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('../lib/index.js')
+require('../lib/cli.js');
