+++ src/cli.ts
@@ -1,186 +1,100 @@
-import * as path from 'path';
-import * as fs from 'fs';
-
+import VodApi from './api';
+import parseCli from './cli-args';
+import { getDesiredBitrate, makeProfile } from './transcode';
+import fs from 'fs';
 import inquirer from 'inquirer';
-import yargs from 'yargs';
-import { hideBin } from 'yargs/helpers';
+import { Asset } from './types/schema';
 
-import { prodApiEndpoint } from './api';
+async function videoNft() {
+	const args = await parseCli();
+	const api = new VodApi(args.apiKey, args.apiEndpoint);
 
-type Camel<T extends string> = T extends `${infer Left}-${infer Right}`
-	? Camel<`${Left}${Capitalize<Right>}`>
-	: T;
+	console.log('1. Requesting upload URL... ');
+	const {
+		url: uploadUrl,
+		asset: { id: assetId },
+		task: importTask
+	} = await api.requestUploadUrl(args.assetName);
+	console.log(`Pending asset with id=${assetId}`);
 
-type CamelKeys<T> = {
-	[K in keyof T as K extends string ? Camel<K> : K]: T[K];
-};
-
-type UnboxPromise<T> = T extends Promise<infer U> ? U : never;
-
-type RawArgs = UnboxPromise<ReturnType<typeof parseRawArgs>>;
-
-export type CliArgs = CamelKeys<{
-	[K in keyof RawArgs]: Exclude<RawArgs[K], undefined>;
-}>;
-
-function parseRawArgs(argv?: string | readonly string[]) {
-	return yargs
-		.command('$0 [filename]', '1-command mint a video NFT')
-		.positional('filename', {
-			describe: 'file to upload as an NFT',
-			type: 'string'
-		})
-		.options({
-			'api-key': {
-				describe: 'API key to use for Livepeer API',
-				type: 'string'
-			},
-			'asset-name': {
-				describe: 'name for the asset created in Livepeer.com API',
-				type: 'string'
-			},
-			'nft-metadata': {
-				describe:
-					'additional JSON metadata to override default generated by Livepeer for the NFT',
-				type: 'string',
-				default: '{}'
-			},
-			'api-endpoint': {
-				describe: 'the endpoint to use for the Livepeer API',
-				type: 'string',
-				default: prodApiEndpoint
-			}
-		})
-		.usage(
-			`
-	Video NFT
-
-  Mint a video NFT in 1 command with Livepeer.
-
-	Usage: video-nft [filename] [options]`
-		)
-		.env('LP_')
-		.help()
-		.parse((argv as any) ?? hideBin(process.argv));
-}
-
-const uuidRegex =
-	/^[0-9a-f]{8}\b-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-\b[0-9a-f]{12}$/;
-
-async function promptMissing(args: RawArgs): Promise<CliArgs> {
-	let { apiKey, filename, assetName, nftMetadata } = args as Record<
-		string,
-		string
-	>;
-	if (!apiKey) {
-		apiKey = await inquirer
-			.prompt({
-				type: 'password',
-				name: 'apiKey',
-				message:
-					'Enter your Livepeer API key (learn more at http://bit.ly/lp-api-key):',
-				mask: '*',
-				validate: apiKey =>
-					uuidRegex.test(apiKey) || 'Not a valid API key'
-			})
-			.then(ans => ans.apiKey as string);
-		console.log(
-			'Tip: You can set the LP_API_KEY environment variable to avoid this prompt.'
-		);
-	}
-	if (!filename) {
-		filename = await inquirer
-			.prompt({
-				type: 'input',
-				name: 'filename',
-				message: 'What file do you want to use?',
-				validate: (value: string) =>
-					fs.existsSync(value) || 'File does not exist'
-			})
-			.then(ans => ans.filename);
-		console.log(
-			'You can also send the filename as an argument to this command.'
-		);
-	}
-	if (!assetName) {
-		assetName = await inquirer
-			.prompt({
-				type: 'input',
-				name: 'assetName',
-				message: `What name do you want to give to your NFT?`,
-				default: path.basename(filename)
-			})
-			.then(ans => ans.assetName);
-	}
-	if (nftMetadata === '{}') {
-		const { shouldEdit } = await inquirer.prompt({
-			type: 'confirm',
-			name: 'shouldEdit',
-			message: `Would you like to customize the NFT metadata?`,
-			default: false
-		});
-		if (shouldEdit) {
-			console.log(
-				' - The `animation_url` and `properties.video` fields will be populated with the exported video URL.'
-			);
-			console.log(' - Set any field to `null` to delete it.');
-			nftMetadata = await inquirer
-				.prompt({
-					type: 'editor',
-					name: 'nftMetadata',
-					message: 'Open text editor:',
-					default: JSON.stringify(
-						{
-							name: assetName,
-							description: `Livepeer video from asset ${JSON.stringify(
-								assetName
-							)}`,
-							image: `ipfs://bafkreidmlgpjoxgvefhid2xjyqjnpmjjmq47yyrcm6ifvoovclty7sm4wm`,
-							properties: {}
-						},
-						null,
-						2
-					),
-					validate: (value: string) => {
-						try {
-							JSON.parse(value);
-							return true;
-						} catch (e) {
-							return `Invalid JSON: ${e}`;
-						}
-					}
-				})
-				.then(ans => ans.nftMetadata);
-		}
-	}
-	return {
-		...args,
-		apiKey,
-		filename,
-		assetName,
-		nftMetadata
-	};
-}
-
-export default async function parseCli(
-	argv?: string | readonly string[]
-): Promise<CliArgs> {
-	const args = await parseRawArgs(argv);
-	if (args.filename && !fs.existsSync(args.filename)) {
-		throw new Error(`File ${args.filename} does not exist`);
-	}
-	if (fs.existsSync(args.nftMetadata)) {
-		args.nftMetadata = fs.readFileSync(args.nftMetadata, 'utf8');
-	}
+	console.log('2. Uploading file...');
+	let file: fs.ReadStream | null = null;
 	try {
-		if (args.nftMetadata != '{}') {
-			const metadata = JSON.parse(args.nftMetadata);
-			console.log(
-				`Using metadata:\n${JSON.stringify(metadata, null, 2)}`
-			);
-		}
-	} catch (e) {
-		throw new Error(`Invalid JSON in nft-metadata: ${e}`);
+		file = fs.createReadStream(args.filename);
+		await api.uploadFile(uploadUrl, file);
+	} finally {
+		file?.close();
 	}
-	return promptMissing(args);
+	await api.waitTask(importTask);
+
+	let asset = await api.getAsset(assetId ?? '');
+	asset = await maybeTranscode(api, asset);
+
+	console.log('3. Starting export... ');
+	let { task: exportTask } = await api.exportAsset(
+		asset.id ?? '',
+		JSON.parse(args.nftMetadata)
+	);
+	console.log(`Created export task with id=${exportTask.id}`);
+	exportTask = await api.waitTask(exportTask);
+
+	const result = exportTask.output?.export?.ipfs;
+	console.log(
+		`4. Export successful! Result: \n${JSON.stringify(result, null, 2)}`
+	);
+
+	console.log(
+		`5. Mint your NFT at:\n` +
+			`https://livepeer.com/mint-nft?tokenUri=${result?.nftMetadataUrl}`
+	);
 }
+
+async function maybeTranscode(api: VodApi, asset: Asset) {
+	const desiredBitrate = await getDesiredBitrate(asset).catch(() => {
+		console.error(
+			`Warning: Asset is larger than OpenSea file limit and can't be transcoded down since it's too large. ` +
+				`It will still be stored in IPFS and referenced in the NFT metadata, so a proper application is still able to play it back. ` +
+				`For more information check http://bit.ly/opensea-file-limit`
+		);
+		return null;
+	});
+	if (!desiredBitrate) {
+		return asset;
+	}
+
+	console.log(
+		`File is too big for OpenSea 100MB limit (learn more at http://bit.ly/opensea-file-limit).`
+	);
+	const { action } = await inquirer.prompt({
+		type: 'list',
+		name: 'action',
+		message: 'What do you want to do?',
+		choices: [
+			{
+				value: 'transcode',
+				name: 'Transcode it to a lower quality so OpenSea is able to preview'
+			},
+			{
+				value: 'ignore',
+				name: 'Mint it as is (should work in any other platform that uses the NFT file)'
+			}
+		]
+	});
+	if (action === 'ignore') {
+		return asset;
+	}
+	const desiredProfile = makeProfile(asset, desiredBitrate);
+	console.log(
+		`3. Transcoding asset to ${desiredProfile.name} at ${Math.round(
+			desiredProfile.bitrate / 1024
+		)} kbps bitrate`
+	);
+	const transcode = await api.transcodeAsset(asset, desiredProfile);
+	await api.waitTask(transcode.task);
+	return transcode.asset;
+}
+
+videoNft().catch(err => {
+	console.error(err);
+	process.exit(1);
+});
