+++ src/transcode.ts
@@ -1,55 +1,35 @@
-import inquirer from 'inquirer';
 import { Asset, FfmpegProfile } from './types/schema';
 
 const openSeaNftSizeLimit = 100_000_000; // 100 MB
 const min720pBitrate = 500_000; // 0.5 Mbps
 const minBitrate = 100_000; // 0.1 Mbps
 
-export async function getDesiredProfile(
-	asset: Asset
-): Promise<FfmpegProfile | null> {
+export async function getDesiredBitrate(
+	asset: Asset,
+	sizeLimit: number = openSeaNftSizeLimit
+): Promise<number | null> {
 	const size = asset.size ?? 0;
-	const videoTrack = asset.videoSpec?.tracks?.find(t => t.type === 'video');
-	const { bitrate, width, height } = videoTrack ?? {};
-	if (size <= openSeaNftSizeLimit || !bitrate || !width || !height) {
+	const bitrate = getVideoTrack(asset)?.bitrate ?? 0;
+	if (size <= sizeLimit || !bitrate) {
 		return null;
 	}
 
 	const audioTrack = asset.videoSpec?.tracks?.find(t => t.type === 'audio');
 	const audioBitrate = audioTrack?.bitrate ?? 0;
 	const desiredBitrate = Math.floor(
-		(bitrate + audioBitrate) * (openSeaNftSizeLimit / size) - audioBitrate
+		(bitrate + audioBitrate) * (sizeLimit / size) - audioBitrate
 	);
 	if (desiredBitrate < minBitrate) {
-		console.error(
-			`Warning: Asset is larger than OpenSea file limit so the video won't playback automatically. ` +
-				`It will still be stored in IPFS and referenced in the NFT metadata, so a proper application is still able to play it back. ` +
-				`For more information check http://bit.ly/opensea-file-limit`
-		);
-		return null;
+		throw new Error('Asset is too large to be downscaled to desired size');
 	}
+	return desiredBitrate;
+}
 
-	console.log(
-		`File is too big for OpenSea 100MB limit (learn more at http://bit.ly/opensea-file-limit).`
-	);
-	const { action } = await inquirer.prompt({
-		type: 'list',
-		name: 'action',
-		message: 'What do you want to do?',
-		choices: [
-			{
-				value: 'transcode',
-				name: 'Transcode it to a lower quality so OpenSea is able to preview'
-			},
-			{
-				value: 'ignore',
-				name: 'Mint it as is (should work in any other platform that uses the NFT file)'
-			}
-		]
-	});
-	if (action === 'ignore') {
-		return null;
-	}
+export function makeProfile(
+	asset: Asset,
+	desiredBitrate: number
+): FfmpegProfile {
+	const { bitrate = 1, width = 0, height = 0 } = getVideoTrack(asset) ?? {};
 
 	// We only change the resolution if the bitrate changes too much. We don't go
 	// below 720p though since the bitrate is the thing that really matters. We
@@ -66,4 +46,8 @@ export async function getDesiredProfile(
 		bitrate: desiredBitrate,
 		fps: 0
 	};
+}
+
+function getVideoTrack(asset: Asset) {
+	return asset.videoSpec?.tracks?.find(t => t.type === 'video');
 }
