+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
 	"include": ["./src/**/*.ts"],
 	"compilerOptions": {
-		"lib": ["es2019", "es2020.promise", "es2020.bigint", "es2020.string"],
+		"lib": ["es2020", "DOM"],
 		"module": "commonjs",
 		"target": "es2019",
